Revert "chore(calculator): rename OrderItemButton -> ~Group"
This reverts commit 9614892967b7783449de642e3bfb539ab1972257.
